Add new methods to controller, new fields to Sensor (SensorAction,SensorType)

+++ skyler/src/main/java/com/skyler/smarthome/server/uicontroller/UIUserController.java
@@ -6,8 +6,10 @@ import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.skyler.smarthome.server.data.UserDao;
@@ -24,32 +26,39 @@ public class UIUserController {
 
 	@RequestMapping(value = "/all", method = RequestMethod.GET, produces = "application/json")
 	public @ResponseBody List<User> getAllUsers() {
-		
 		List<User> users = userDao.getAllUser();
 		return users;
 	}
 
 	@RequestMapping(value = "/new", method = RequestMethod.POST)
-	public void createUser(User user) {
-		userDao.createNewUser(user);
-
+	public void createUser(@RequestBody User user) {
+		if (user != null) {
+			boolean result = userDao.createNewUser(user);
+		}
 	};
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = "application/json")
 	public @ResponseBody User getUserById(@PathVariable int id) {
-		User user = userDao.getUserById(id);
-		return user;
-
+		if (id > 0) {
+			User user = userDao.getUserById(id);
+			return user;
+		} else {
+			return null;
+		}
 	};
 
-	@RequestMapping(value = "/update", method = RequestMethod.POST)
-	public void updateUser(User user) {
-		userDao.updateUser(user);
+	@RequestMapping(value = "/update/{id}", method = RequestMethod.POST)
+	public void updateUser(@PathVariable int id, @RequestParam String userField, @RequestParam String newParam) {
+		if (id > 0 && userField != null && newParam != null) {
+			boolean result = userDao.updateUser(id, userField, newParam);
+		}
 	};
 
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.POST)
 	public void deleteUserById(@PathVariable int id) {
-		userDao.deleteUserById(id);
+		if (id > 0) {
+			boolean result = userDao.deleteUserById(id);
+		}
 	};
 
 }
